refactor(dashboard): separate rows from viewport projection
