+++ packages/dashboard-core/src/selectors/dashboardViewport.ts
@@ -1,56 +1,22 @@
 import type { ProjectId, ProjectView, StationSnapshot } from "@station/contracts";
 import { clampDashboardScrollOffset, dashboardBodyRows } from "../components/Dashboard/layout.js";
-import { worktreeRowVisibleFields } from "../components/WorktreeRow/rowInput.js";
-import type {
-  FailedCreateSessionRow,
-  PendingCreateSessionRow,
-  PendingRemoveWorktreeRow,
-  PendingStartAgentRow,
-} from "../state/localRows.js";
 import type { TuiScreen, TuiViewState } from "../state/types.js";
 import {
-  type DashboardPersistentFilterCandidate,
   type DashboardPersistentFilterProjection,
   type DashboardPersistentFilterProjectMatch,
-  type DashboardPersistentFilterRowMatch,
-  type DashboardPersistentFilterVisibleFields,
   selectDashboardPersistentFilter,
 } from "./dashboardPersistentFilter.js";
-import { matchesDashboardOptimisticSearch } from "./dashboardSearchProjection.js";
 import {
-  type DashboardSessionRow,
-  type KeyedChoice,
-  keyChoices,
-  selectProjectGroups,
-  sessionRowDisplayTitle,
-} from "./selectors.js";
-
-export type DashboardCreateSessionLocalRow =
-  | ({ status: "pending" } & PendingCreateSessionRow)
-  | ({ status: "failed" } & FailedCreateSessionRow);
-
-export type DashboardRowPresentation = DashboardPersistentFilterVisibleFields;
-
-type DashboardSessionViewportItem = {
-  type: "session";
-  id: string;
-  row: DashboardSessionRow;
-  displayTitle: string;
-  presentation: DashboardRowPresentation;
-  pendingRemove?: PendingRemoveWorktreeRow;
-  pendingStart?: PendingStartAgentRow;
-  persistentFilterMatch?: DashboardPersistentFilterRowMatch;
-};
-
-type DashboardCreateLocalViewportItem = {
-  type: "createLocalRow";
-  id: string;
-  row: DashboardCreateSessionLocalRow;
-  presentation: DashboardRowPresentation;
-  persistentFilterMatch?: DashboardPersistentFilterRowMatch;
-};
-
-type DashboardRowViewportItem = DashboardSessionViewportItem | DashboardCreateLocalViewportItem;
+  type DashboardCreateLocalItem,
+  type DashboardProjectRowGroup,
+  type DashboardRowItem,
+  type DashboardSessionItem,
+  persistentFilterCandidateForDashboardRow,
+  selectDashboardProjectRowGroups,
+  withDashboardRowFilterMatch,
+} from "./dashboardProjectRows.js";
+import type { DashboardSessionRow } from "./dashboardSessionRows.js";
+import { type KeyedChoice, keyChoices } from "./selectors.js";
 
 export type DashboardViewportItem =
   | {
@@ -71,8 +37,8 @@ export type DashboardViewportItem =
       id: string;
       project: ProjectView;
     }
-  | DashboardSessionViewportItem
-  | DashboardCreateLocalViewportItem;
+  | DashboardSessionItem
+  | DashboardCreateLocalItem;
 
 export type DashboardViewport = {
   bodyRows: number;
@@ -145,10 +111,6 @@ export function selectDashboardViewport(
   return viewport;
 }
 
-function countSessionRows(items: readonly DashboardViewportItem[]): number {
-  return items.filter((item) => item.type === "session" || item.type === "createLocalRow").length;
-}
-
 export function selectDashboardItems(
   snapshot: StationSnapshot,
   state: DashboardViewportState,
@@ -161,12 +123,6 @@ export function selectDashboardItems(
   ).items;
 }
 
-type DashboardItemGroup = {
-  project: ProjectView;
-  collapsed: boolean;
-  children: DashboardRowViewportItem[];
-};
-
 function selectDashboardItemsProjection(
   snapshot: StationSnapshot,
   state: TuiViewState,
@@ -177,11 +133,11 @@ function selectDashboardItemsProjection(
 } {
   const persistentFilterSelected =
     activeScreen.name === "persistentFilter" || state.persistentFilter !== undefined;
-  const groups = selectDashboardItemGroups(snapshot, state, {
+  const groups = selectDashboardProjectRowGroups(snapshot, state, {
     applyLegacySearch: !persistentFilterSelected,
   });
   const persistentFilter = selectDashboardPersistentFilter({
-    candidates: groups.flatMap((group) => group.children.map(persistentFilterCandidate)),
+    candidates: groups.flatMap((group) => group.rows.map(persistentFilterCandidateForDashboardRow)),
     projects: groups.map((group) => ({
       projectId: group.project.id,
       projectLabel: group.project.label,
@@ -189,123 +145,68 @@ function selectDashboardItemsProjection(
     screen: activeScreen,
     ...(state.persistentFilter === undefined ? {} : { applied: state.persistentFilter }),
   });
-
-  if (persistentFilter === undefined) {
-    return { items: flattenDashboardGroups(groups), persistentFilter };
-  }
-  if (persistentFilter.source === "draft" || persistentFilter.query.length === 0) {
-    return {
-      items: flattenDashboardGroups(groups, persistentFilter),
-      persistentFilter,
-    };
-  }
   return {
-    items: flattenAppliedDashboardGroups(groups, persistentFilter),
+    items: flattenDashboardGroups(groups, persistentFilter),
     persistentFilter,
   };
 }
 
-function selectDashboardItemGroups(
-  snapshot: StationSnapshot,
-  state: TuiViewState,
-  options: { applyLegacySearch: boolean },
-): DashboardItemGroup[] {
-  const localRows = visibleCreateSessionLocalRows(snapshot, state);
-  return selectProjectGroups(snapshot, state, {
-    includeCollapsedRows: true,
-    applySearch: options.applyLegacySearch,
-  }).map((group) => {
-    const projectLocalRows = localRows
-      .filter((row) => row.projectId === group.project.id)
-      .filter(
-        (row) =>
-          !options.applyLegacySearch ||
-          matchesDashboardOptimisticSearch(
-            {
-              title: row.title,
-              branch: row.branch,
-              projectLabel: group.project.label,
-              pendingHarnessProvider: row.status === "pending" ? row.harnessProvider : undefined,
-            },
-            state.searchQuery,
-          ),
-      );
-    return {
-      project: group.project,
-      collapsed: group.collapsed,
-      children: mergeRowsAndCreateSessionLocalRows(group.rows, projectLocalRows, state).map(
-        (row) =>
-          row.type === "session"
-            ? sessionViewportItem(row.row, state)
-            : createLocalViewportItem(row.row),
-      ),
-    };
-  });
-}
-
 function flattenDashboardGroups(
-  groups: readonly DashboardItemGroup[],
+  groups: readonly DashboardProjectRowGroup[],
   projection?: DashboardPersistentFilterProjection,
 ): DashboardViewportItem[] {
-  return groups.flatMap((group, index) => {
-    const items: DashboardViewportItem[] = [];
-    if (index > 0) {
-      items.push({
-        type: "projectGap",
-        id: `gap:${group.project.id}`,
-        projectId: group.project.id,
-      });
-    }
-    items.push(projectHeaderItem(group, projection));
-    if (group.collapsed) {
-      return items;
-    }
-    if (group.children.length === 0) {
-      items.push(emptyProjectItem(group.project));
-      return items;
-    }
-    items.push(
-      ...group.children.map((child) =>
-        projection === undefined ? child : attachPersistentFilterRowMatch(child, projection),
-      ),
-    );
-    return items;
-  });
+  const applied = projection?.source === "applied" && projection.query.length > 0;
+  const visibleGroups = applied
+    ? groups.filter((group) => projection.projects.get(group.project.id)?.matched)
+    : groups;
+  return visibleGroups.flatMap((group, index) => [
+    ...(index === 0 ? [] : [projectGapItem(group.project.id)]),
+    ...projectGroupItems(group, projection, applied),
+  ]);
 }
 
-function flattenAppliedDashboardGroups(
-  groups: readonly DashboardItemGroup[],
-  projection: DashboardPersistentFilterProjection,
+function projectGroupItems(
+  group: DashboardProjectRowGroup,
+  projection: DashboardPersistentFilterProjection | undefined,
+  applied: boolean,
 ): DashboardViewportItem[] {
-  const retained = groups.filter((group) => projection.projects.get(group.project.id)?.matched);
-  return retained.flatMap((group, index) => {
-    const items: DashboardViewportItem[] = [];
-    if (index > 0) {
-      items.push({
-        type: "projectGap",
-        id: `gap:${group.project.id}`,
-        projectId: group.project.id,
-      });
-    }
-    items.push(projectHeaderItem(group, projection));
-    const matchingChildren = group.children.filter(
-      (child) => projection.rows.get(child.id)?.matched === true,
-    );
-    if (matchingChildren.length === 0) {
-      if (group.children.length === 0) {
-        items.push(emptyProjectItem(group.project));
-      }
-      return items;
-    }
+  const items: DashboardViewportItem[] = [projectHeaderItem(group, projection)];
+  const rows = visibleProjectRows(group, projection, applied);
+  if (rows.length > 0) {
     items.push(
-      ...matchingChildren.map((child) => attachPersistentFilterRowMatch(child, projection)),
+      ...rows.map((row) =>
+        projection === undefined ? row : withDashboardRowFilterMatch(row, projection),
+      ),
     );
-    return items;
-  });
+  } else if (group.rows.length === 0 && (!group.collapsed || applied)) {
+    items.push(emptyProjectItem(group.project));
+  }
+  return items;
+}
+
+function visibleProjectRows(
+  group: DashboardProjectRowGroup,
+  projection: DashboardPersistentFilterProjection | undefined,
+  applied: boolean,
+): readonly DashboardRowItem[] {
+  if (applied && projection !== undefined) {
+    return group.rows.filter((row) => projection.rows.get(row.id)?.matched === true);
+  }
+  return group.collapsed ? [] : group.rows;
+}
+
+function projectGapItem(
+  projectId: ProjectId,
+): Extract<DashboardViewportItem, { type: "projectGap" }> {
+  return {
+    type: "projectGap",
+    id: `gap:${projectId}`,
+    projectId,
+  };
 }
 
 function projectHeaderItem(
-  group: DashboardItemGroup,
+  group: DashboardProjectRowGroup,
   projection?: DashboardPersistentFilterProjection,
 ): Extract<DashboardViewportItem, { type: "projectHeader" }> {
   const item: Extract<DashboardViewportItem, { type: "projectHeader" }> = {
@@ -331,99 +232,6 @@ function emptyProjectItem(
   };
 }
 
-function sessionViewportItem(
-  row: DashboardSessionRow,
-  state: TuiViewState,
-): DashboardSessionViewportItem {
-  const displayTitle = sessionRowDisplayTitle(row, state.localRows);
-  const pendingRemove = state.localRows.pendingRemove.find(
-    (localRow) => localRow.worktreeId === row.worktree.id,
-  );
-  const pendingStart = state.localRows.pendingStart.find(
-    (localRow) => localRow.worktreeId === row.worktree.id,
-  );
-  const item: DashboardSessionViewportItem = {
-    type: "session",
-    id: `session:${row.id}`,
-    row,
-    displayTitle,
-    presentation: sessionRowPresentation(row, displayTitle, pendingRemove, pendingStart),
-  };
-  if (pendingRemove !== undefined) {
-    item.pendingRemove = pendingRemove;
-  }
-  if (pendingStart !== undefined) {
-    item.pendingStart = pendingStart;
-  }
-  return item;
-}
-
-function createLocalViewportItem(
-  row: DashboardCreateSessionLocalRow,
-): DashboardCreateLocalViewportItem {
-  return {
-    type: "createLocalRow",
-    id: `create:${row.localId}`,
-    row,
-    presentation: createSessionRowPresentation(row),
-  };
-}
-
-function persistentFilterCandidate(
-  item: DashboardRowViewportItem,
-): DashboardPersistentFilterCandidate {
-  return {
-    kind: item.type === "session" ? "session" : "optimistic",
-    id: item.id,
-    projectId: item.type === "session" ? item.row.worktree.projectId : item.row.projectId,
-    visibleFields: item.presentation,
-  };
-}
-
-function attachPersistentFilterRowMatch(
-  item: DashboardRowViewportItem,
-  projection: DashboardPersistentFilterProjection,
-): DashboardRowViewportItem {
-  const match = projection.rows.get(item.id);
-  return match === undefined ? item : { ...item, persistentFilterMatch: match };
-}
-
-function sessionRowPresentation(
-  row: DashboardSessionRow,
-  displayTitle: string,
-  pendingRemove: PendingRemoveWorktreeRow | undefined,
-  pendingStart: PendingStartAgentRow | undefined,
-): DashboardRowPresentation {
-  if (pendingRemove !== undefined) {
-    return { title: displayTitle, activity: "removing session..." };
-  }
-  if (pendingStart !== undefined) {
-    return {
-      title: displayTitle,
-      activity: pendingStart.operation === "resumeAgent" ? "resuming..." : "starting...",
-    };
-  }
-  const visibleFields = worktreeRowVisibleFields(row.presentation, displayTitle);
-  return {
-    title: visibleFields.title,
-    agent: visibleFields.agent,
-    activity: visibleFields.activity,
-  };
-}
-
-function createSessionRowPresentation(
-  row: DashboardCreateSessionLocalRow,
-): DashboardRowPresentation {
-  if (row.status === "failed") {
-    return { title: row.title, activity: row.error.message };
-  }
-  return {
-    title: row.title,
-    agent: row.harnessProvider ?? "",
-    activity: "starting session...",
-  };
-}
-
 function displaySessionRowsFromItems(
   items: readonly DashboardViewportItem[],
 ): DashboardSessionRow[] {
@@ -432,77 +240,6 @@ function displaySessionRowsFromItems(
   );
 }
 
-type GroupDashboardRow =
-  | {
-      type: "session";
-      row: DashboardSessionRow;
-    }
-  | {
-      type: "createLocalRow";
-      row: DashboardCreateSessionLocalRow;
-    };
-
-function visibleCreateSessionLocalRows(
-  snapshot: StationSnapshot,
-  state: TuiViewState,
-): DashboardCreateSessionLocalRow[] {
-  const rowsById = new Map(snapshot.rows.map((row) => [row.id, row]));
-  const realRows = new Set(
-    snapshot.sessions.flatMap((session) => {
-      const row = rowsById.get(session.worktreeId);
-      return row === undefined ? [] : [`${session.projectId}\u0000${row.branch}`];
-    }),
-  );
-  return [
-    ...state.localRows.pendingCreate.flatMap((row) =>
-      realRows.has(`${row.projectId}\u0000${row.branch}`)
-        ? []
-        : [{ ...row, status: "pending" as const }],
-    ),
-    ...state.localRows.failedCreate.map((row) => ({
-      ...row,
-      status: "failed" as const,
-    })),
-  ];
-}
-
-function mergeRowsAndCreateSessionLocalRows(
-  rows: readonly DashboardSessionRow[],
-  localRows: readonly DashboardCreateSessionLocalRow[],
-  state: TuiViewState,
-): GroupDashboardRow[] {
-  return [
-    ...rows.map((row) => ({ type: "session" as const, row })),
-    ...localRows.map((row) => ({ type: "createLocalRow" as const, row })),
-  ].sort((left, right) => compareDashboardRows(left, right, state));
-}
-
-function compareDashboardRows(
-  left: GroupDashboardRow,
-  right: GroupDashboardRow,
-  state: TuiViewState,
-): number {
-  const titleOrder = rowTitle(left, state).localeCompare(rowTitle(right, state));
-  if (titleOrder !== 0) return titleOrder;
-  const branchOrder = rowBranch(left).localeCompare(rowBranch(right));
-  if (branchOrder !== 0) return branchOrder;
-  if (left.type !== right.type) {
-    return left.type === "session" ? -1 : 1;
-  }
-  return rowId(left).localeCompare(rowId(right));
-}
-
-function rowTitle(row: GroupDashboardRow, state: TuiViewState): string {
-  if (row.type === "createLocalRow") {
-    return row.row.title;
-  }
-  return sessionRowDisplayTitle(row.row, state.localRows);
-}
-
-function rowBranch(row: GroupDashboardRow): string {
-  return row.type === "session" ? row.row.worktree.branch : row.row.branch;
-}
-
-function rowId(row: GroupDashboardRow): string {
-  return row.type === "session" ? row.row.id : row.row.localId;
+function countSessionRows(items: readonly DashboardViewportItem[]): number {
+  return items.filter((item) => item.type === "session" || item.type === "createLocalRow").length;
 }
