+++ packages/dashboard-core/src/selectors/fleetSummary.ts
@@ -1,6 +1,6 @@
 import type { StationSnapshot } from "@station/contracts";
 import { isReadyToRead } from "../components/WorktreeRow/rowInput.js";
-import { selectDashboardSessionRows } from "./selectors.js";
+import { selectDashboardSessionRows } from "./dashboardSessionRows.js";
 
 // Fleet triage counts derived client-side: the observer's snapshot.counts carry
 // only working/idle/attention/unknown and fold "ready" into idle, so the fleet
